Updated to use floats instead of ints

+++ revision_1/alt_c_to_f_converter.c
@@ -4,7 +4,7 @@
     for fahr = 0, 20, ..., 300 */
 int main(void)
 {
-    int fahr, celsius;
+    float fahr, celsius;
     int lower, upper, step;
 
     lower = 0;
@@ -14,8 +14,8 @@ int main(void)
     fahr = lower;
     while (fahr <= upper)
     {
-        celsius = (5 * (fahr - 32)) / 9;
-        printf("%d\t%d\n", fahr, celsius);
+        celsius = (5.0 * (fahr - 32.0)) / 9.0;
+        printf("%3.0f\t%6.1f\n", fahr, celsius);
         fahr = fahr + step;
     }
 }
